+++ crates/mongodb-connector/src/mongo_connector.rs
@@ -2,6 +2,7 @@ use std::path::Path;
 
 use anyhow::anyhow;
 use async_trait::async_trait;
+use bytes::Bytes;
 use configuration::Configuration;
 use mongodb_agent_common::{
     explain::explain_query, health::check_health, interface_types::MongoConfig,
@@ -9,7 +10,8 @@ use mongodb_agent_common::{
 };
 use ndc_sdk::{
     connector::{
-        Connector, ConnectorSetup, ExplainError, FetchMetricsError, HealthError, InitializationError, MutationError, ParseError, QueryError, SchemaError
+        Connector, ConnectorSetup, ExplainError, FetchMetricsError, HealthError,
+        InitializationError, MutationError, ParseError, QueryError, SchemaError,
     },
     json_response::JsonResponse,
     models::{
@@ -23,7 +25,8 @@ use crate::{
     api_type_conversions::{
         v2_to_v3_explain_response, v2_to_v3_query_response, v3_to_v2_query_request, QueryContext,
     },
-    error_mapping::{mongo_agent_error_to_explain_error, mongo_agent_error_to_query_error}, schema,
+    error_mapping::{mongo_agent_error_to_explain_error, mongo_agent_error_to_query_error},
+    schema,
 };
 use crate::{capabilities::mongo_capabilities_response, mutation::handle_mutation_request};
 
@@ -159,16 +162,18 @@ impl Connector for MongoConnector {
             .await
             .map_err(mongo_agent_error_to_query_error)?;
 
-        // TODO: This requires parsing and reserializing the response from MongoDB. We can avoid
-        // this by passing a response format enum to the query pipeline builder that will format
-        // responses differently for v3 vs v2. MVC-7
-        let response = response_json
-            .into_value()
-            .map_err(|e| QueryError::Other(Box::new(e)))?;
-
-        // TODO: If we are able to push v3 response formatting to the MongoDB aggregation pipeline
-        // then we can switch to using `map_unserialized` here to avoid  deserializing and
-        // reserializing the response. MVC-7
-        Ok(v2_to_v3_query_response(response).into())
+        match response_json {
+            dc_api::JsonResponse::Value(v2_response) => {
+                Ok(JsonResponse::Value(v2_to_v3_query_response(v2_response)))
+            }
+            dc_api::JsonResponse::Serialized(bytes) => {
+                let v2_value: serde_json::Value = serde_json::de::from_slice(&bytes)
+                    .map_err(|e| QueryError::Other(Box::new(e)))?;
+                let v3_bytes: Bytes = serde_json::to_vec(&vec![v2_value])
+                    .map_err(|e| QueryError::Other(Box::new(e)))?
+                    .into();
+                Ok(JsonResponse::Serialized(v3_bytes))
+            }
+        }
     }
 }